requirements.txt
django==5.1.5
djangorestframework==3.15.2
psycopg2==2.9.10
django-split-settings==1.3.2